Migrate bookmarks to app router
